requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
google-generativeai
langchain
langchain-google-genai
